+++ src/tunnels/gtp_tunnel.cpp
@@ -12,7 +12,6 @@
 typedef std::map<uint32_t, GTPU *, std::less<uint32_t> > gtpu_flow_map_t;
 typedef gtpu_flow_map_t::const_iterator gtpu_flow_map_iter_t;
 typedef gtpu_flow_map_t::iterator gtpu_flow_map_iter_no_const_t;
-gtpu_flow_map_t mapOfTeid_v4;
 
 /**
  *      Bits
@@ -386,132 +385,6 @@ int GTPU::Adjust(rte_mbuf * pkt, u_int8_t queue ) {
 	return 0;
 }
 
-uint16_t Tunnel::RxCallback(uint16_t port, uint16_t queue,
-		struct rte_mbuf *pkts[], uint16_t nb_pkts, uint16_t max_pkts,
-		void *user_param) {
-
-    GTPU t;
-    u_int8_t i = 0;
-
-	if (0 == nb_pkts)
-	    return nb_pkts;
-
-    rte_mbuf *m = NULL;
-    /* Prefetch first 3 packets */
-    for (i = 0; i < 3 && i < nb_pkts; i++) {
-        rte_prefetch0(rte_pktmbuf_mtod(
-                pkts[i], void *));
-    }
-
-    /*Do batch processing for next set of 3 packets each time */
-    for (i = 0; i < (nb_pkts -3 ); i++) {
-        rte_prefetch0(rte_pktmbuf_mtod(pkts[i + 3], void *));
-     	m = pkts[i];
-        t.Adjust(m,queue);
-	}
-
-    /*Process the remaining packets*/
-    for (; i < nb_pkts ; i++) {
-        m = pkts[i];
-        t.Adjust(m,queue);
-    }
-
-	return nb_pkts;
-}
-
-uint16_t Tunnel::TxCallback(uint16_t port, uint16_t queue,
-		struct rte_mbuf *pkts[], uint16_t nb_pkts, void *user_param) {
-
-	if (!CGlobalInfo::m_options.m_ip_cfg[port].is_gtp_enabled()) {
-		return nb_pkts;
-	}
-
-    u_int8_t i = 0 ; 
-    struct rte_ipv4_hdr *iph __attribute__((unused))  = NULL;
-    Tunnel *tunnel;
-   
-
-    /*Get the Tunnel instance based on the subscriber IP address*/
-
-    u_int8_t is_ipv4 = 1;
-
-     /* Prefetch first 3 packets */
-     for (i = 0; i < 3 && i < nb_pkts; i++) {
-         rte_prefetch0(rte_pktmbuf_mtod(pkts[i], void *));
-     }
-
-     /*Do batch processing for next set of 3 packets each time */
-	for (i=0; i < (nb_pkts -3 ); i++) {
-        rte_prefetch0(rte_pktmbuf_mtod(pkts[i + 3], void *));
-		rte_mbuf * m = pkts[i];
-        iph = rte_pktmbuf_mtod_offset(m, struct rte_ipv4_hdr *, m->l2_len);
-	    tunnel = (Tunnel *)m->dynfield1[0];
-        if(tunnel) {
-	        tunnel->Prepend(m,is_ipv4, queue);
-        }
-	}
-
-    /*Process the remaining packets*/
-    for (; i < nb_pkts; i++) {
-        rte_mbuf * m = pkts[i];
-        iph = rte_pktmbuf_mtod_offset(m, struct rte_ipv4_hdr *, m->l2_len);
-	    tunnel = (Tunnel *)m->dynfield1[0];
-        if(tunnel) {
-          tunnel->Prepend(m,is_ipv4, queue);
-        }
-    }
-    return nb_pkts;
- }
-
-void Tunnel::InstallRxCallback(uint16_t port, uint16_t queue) {
-
-
-    if (!CGlobalInfo::m_options.m_ip_cfg[port].is_gtp_enabled()) {
-        printf("Tx Tunnel disabled on %u:%u...But doing Rx callback", port, 0);
-    }
-
-	if (NULL == rte_eth_add_rx_callback(port, queue, Tunnel::RxCallback, NULL)) {
-		printf("failed to install RX callback at %u:%u.", port, queue);
-		return;
-	}
-	printf("installed RX callback at %u:%u.", port, queue);
-}
-
-void Tunnel::InstallTxCallback(uint16_t port, uint16_t queue) {
-
-    if (!CGlobalInfo::m_options.m_ip_cfg[port].is_gtp_enabled()) {
-        printf("Tx Tunnel disabled on %u:%u.", port, queue);
-        return ;
-    }
-
-	if (NULL == rte_eth_add_tx_callback(port, queue, Tunnel::TxCallback, NULL)) {
-		printf("failed to install TX callback at %u:%u.", port, 0);
-		return;
-	}
-	printf("installed TX callback at %u:%u.", port, queue);
-}
-
-int Tunnel::Parse(YAML::Node const& node, std::vector<Tunnel *> & tunnels)
-{
-	if (node.FindValue("type")) {
-        std::string type;
-        node["type"] >> type;
-		if (type == "gtp") {
-            const YAML::Node& tep_info = node["tep"];
-            for(unsigned i=0;i<tep_info.size();i++) {
-                const YAML::Node & tep_node = tep_info;
-                GTPU::Parse(tep_node[i],tunnels);
-            }
-       }
-       else {
-           /*For now Only GTP tunnel is supported */
-           ASSERT_MSG(false, "Unknown tunnel type. only Supports 'gtp' tunnel");
-       }
-    }
-	return 0;
-
-}
-
 extern int my_inet_pton4(const char *src, unsigned char *dst);
 inline bool ParseIPv4(YAML::Node const& node, uint32_t * ip) {
     std::string ip_str;
@@ -584,54 +457,3 @@ GTPU::GTPU(uint32_t teid, uint8_t src_ipv6[16], uint8_t dst_ipv6[16])
     m_ipv6_set = 1;
 
 }
-
-
-int GTPU::Parse(YAML::Node const& node, std::vector<Tunnel *> & tunnels) 
-{
-    uint32_t teid, src_ip,dst_ip, key;
-    GTPU *new_elem;
-
-    if (!utl_yaml_read_uint32(node, "teid", teid)) {
-	    ASSERT_MSG(false, "Can't find 'tun_id' in GTP tunnel.");
-	}
-
-	if (!node.FindValue("src_ip")) {
-	    ASSERT_MSG(false, "Can't find 'tun_src' in GTP tunnel.");
-	}
-    
-    if (!utl_yaml_read_ip_addr(node, "dst_ip", dst_ip)) {
-	    ASSERT_MSG(false, "Can't find 'dst_ip' in GTP tunnel.");
-	}
-
-    if (!utl_yaml_read_ip_addr(node, "src_ip", src_ip)) {
-	    ASSERT_MSG(false, "Can't find 'src_ip' in GTP tunnel.");
-	}
-
-    if (!utl_yaml_read_ip_addr(node, "key", key)) {
-	    ASSERT_MSG(false, "Can't find 'Key(UE_IP)' in GTP tunnel.");
-    }
-
-    new_elem = new GTPU(teid,PKT_HTONL(src_ip),PKT_HTONL(dst_ip));
-    tunnels.push_back(new_elem);
-
-    std::pair<uint32_t, GTPU *> pair = std::pair<uint32_t, GTPU *>(PKT_HTONL(key), new_elem);
-    mapOfTeid_v4.insert(pair);
-
-    if ( node.FindValue("src_ipv6") ) {
-        uint8_t src_ipv6[8];
-        uint8_t dst_ipv6[8];
-        parse_ipv6(node, "src_ipv6", (unsigned char *)&src_ipv6);
-
-        if ( node.FindValue("dst_ipv6") ) 
-            parse_ipv6(node, "dst_ipv6", (unsigned char *)&dst_ipv6);
-        else
-            ASSERT_MSG(false, "Can't find 'dst_ipv6 ' in GTP IPv6 tunnel.");
-
-        if (!utl_yaml_read_uint32(node, "teid_v6", teid)) {
-            ASSERT_MSG(false, "Can't find 'tun_id' in GTP tunnel Ipv6.");
-        }
-        tunnels.push_back(new GTPU(teid, src_ipv6, dst_ipv6 ) );
-   }
-   return 0;
-}
-
